Added bitsPerPixel support to color convert and fixed some issues that showed up when chaining plugins together
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/ADCore/current@20904 ececa60f-9f63-0410-97c7-ea05888eb260

+++ ADApp/pluginSrc/NDPluginROI.cpp
@@ -220,7 +220,7 @@ void NDPluginROI::processCallbacks(NDArray *pArray)
 	/* Calculate ROI bitsPerElement */
     int		bitsPerPixel	=	arrayInfo.bitsPerElement;
 	size_t	binFactor = 1;
-	for ( int iDim=0; iDim < pArray->ndims; iDim++ )
+	for ( size_t iDim=0; iDim < pArray->ndims; iDim++ )
 		binFactor	*= dims[iDim].binning;
 	if ( binFactor != 1 )
 		bitsPerPixel	+= static_cast<int>( ceil( log2( binFactor ) ) );
